+++ frontend/src/components/ui/Ribbon.jsx
@@ -4,13 +4,18 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import styles from './Ribbon.module.less';
 
-const Ribbon = ({ children, color, placement = 'end', show, text }) => (
-  <Badge.Ribbon color={color} placement={placement} text={text}>
-    <div className={classNames(styles.ribbon, { [styles.hidden]: !show })}>
-      {children}
-    </div>
-  </Badge.Ribbon>
-);
+const Ribbon = ({ children, color, placement = 'end', show, text }) => {
+  const safeShow = text && show;
+  return (
+    <Badge.Ribbon color={color} placement={placement} text={text}>
+      <div
+        className={classNames(styles.ribbon, { [styles.hidden]: !safeShow })}
+      >
+        {children}
+      </div>
+    </Badge.Ribbon>
+  );
+};
 
 Ribbon.propTypes = {
   children: PropTypes.node,