+++ frontend/src/styles/theme.js
@@ -1,20 +1,18 @@
 /** Try it out at https://ant.design/theme-editor */
 
 const theme = {
-  colorPrimary: '#edb048',
-  colorInfo: '#ff5d5f',
-  colorSuccess: '#49aa19', // Dark mode
-  colorWarning: '#d89614', // Dark mode
-  colorError: '#a61d24', // Dark mode
-  colorTextBase: '#ffffffdd',
-  colorBgBase: '#000000',
-  colorBgLayout: '#222222',
-  colorBgLayoutDarker: '#181818',
+  colorPrimary: '#6873e7',
+  colorInfo: '#D1B75C',
+  colorSuccess: '#52c41a',
+  colorWarning: '#faad14',
+  colorError: '#ff4d4f',
+  colorTextBase: '#fdfeff',
+  colorBgBase: '#201c24',
   baseBorderRadius: '6px',
   boxShadow:
-    '0 1px 2px 0 rgba(0, 0, 0, 0.03), 0 1px 6px -1px rgba(0, 0, 0, 0.02), 0 2px 4px 0 rgba(0, 0, 0, 0.02)',
+    '0 1px 2px 0 rgba(32, 28, 36 0.03), 0 1px 6px -1px rgba(32, 28, 36 0.02), 0 2px 4px 0 rgba(32, 28, 36 0.02)',
   boxShadowSecondary:
-    '0 6px 16px 0 rgba(0, 0, 0, 0.08),0 3px 6px -4px rgba(0, 0, 0, 0.12),0 9px 28px 8px rgba(0, 0, 0, 0.05)',
+    '0 6px 16px 0 rgba(32, 28, 36 0.08),0 3px 6px -4px rgba(32, 28, 36 0.12),0 9px 28px 8px rgba(32, 28, 36 0.05)',
 };
 
 export default theme;